[xla][gpu] Fixed two problems in p2p-schedule-preparation.
Previously, we record the address of a P2PGroupNode for a complement-group that
forms a cycle with a group. This address can be changed as P2PGroupNode is an
entity of a map. Fixing this problem by recording the channel ID of the
complement-group instead.

Make sure that a group and its complement-group are both annotated with
a stream for pipelining.

PiperOrigin-RevId: 624219900