dump: Respect dumping configuration in DumpProtobufToFile.
Dumping currently does not happen when disabled from this function because
the IsDirectory check will fail on the empty string, but that's both
unclear and unexpected, so let's be more precise.

While here, add a bunch of missing includes/build deps.

New unit tests are added around this functionality both for this method and for
DumpToFileInDir.

PiperOrigin-RevId: 642643659