[XLA:MSA] Sync copy replacement
This change adds a feature to MSA that replaces synchronous copies (between default and alternate memory) with asynchronous copies, when possible. The feature is disabled by default for now.

The feature is implemented by creating a closure of intervals for each buffer in sorted_buffer_intervals. Each closure has an entrance interval and all intervals of a closure are connected to the entrance interval through at least one sync copy instruction that we are trying to replace. The closure is then processed in a loop, and the allocation values for the closure are created. If all allocation values of the closure can be allocated successfully, and the sync copy instruction replacement correctness conditions are met, the sync copy will be replaced with an async copy instruction. If not, the sync copy instruction is kept as is, and will be ignored in replacement efforts of future closures. For a sync copy to be consider replaceable, beside not being in the ignored list because of the past failures, it must not change the layout between its input and output, it must not have a preassigned memory space for its operand or output, and its operand value must only show up in one position.

In order to replace a copy, we create segments from the AllocationValue to the uses of its copy uses that are considered for replacement.

FUTURE_COPYBARA_INTEGRATE_REVIEW=https://github.com/tensorflow/tensorflow/pull/63359 from Notheisz57:Noth_PyInterpreterWrapper_SegFault_Fix 7ad5a7612918285ddca49633d63c462262bd3dc9
PiperOrigin-RevId: 643097604