Fix breakage to TF Serving
TF Serving has not been able to release due to a breakage in one
of its dependencies.

mkl_sparse_matrix_matmul_op began using CSRMatMulOp in
https://github.com/tensorflow/tensorflow/pull/63030. CSRMatMulOp
uses extensive templating but the templating specialization was
split between the header and cc files. This cl moves it entirely into
the header file.

For some reason, the original change did not cause breakages to
the main TF release but it does break TF serving when it is included.

PiperOrigin-RevId: 648831344