[XLA:CPU] Add runtime check if `get-dimension-size` was rewritten.
In the current runtime, `get-dimension-size` op emitting is not supported by design, normally it is rewritten by `DynamicPadder` HLO pass and never reaches the emit phase. This CL adds a runtime check if this op was actually rewritten and returns an explicit message if it wasn't (instead of cryptic message that kGetDimensionSize opcode was not handled in elemental IR emitter).

A relevant test case was added to `get_dimension_size_test.cc`.

Besides the runtime check, this CL also turns on `get_dimension_size` tests for the new runtime. Because this op is rewritten as already supported ops, we get support 'for free' in the new runtime.

PiperOrigin-RevId: 649397026