[XLA:CPU] Add runtime check if `set-dimension-size` was rewritten.
In the current runtime, `set-dimension-size` op emitting is not supported by design, normally it is rewritten by `DynamicPadder` HLO pass and never reaches the emit phase. This CL adds a runtime check if this op was actually rewritten and returns an explicit message if it wasn't (instead of cryptic message that kSetDimensionSize opcode was not handled in elemental IR emitter).

Also added unit tests, covering existing and new functionality: `set_dimension_size_test.cc`.

Can't turn these tests for new runtime at this moment, because `set-dimension-size` op is rewritten as `SliceToDynamic` custom call (which is not supported yet).

PiperOrigin-RevId: 649630355