[XLA:GPU] Make `SoftmaxRewriterTriton` use the new `IsTritonSupportedInstruction`, as opposed to the one in `legacy_triton::`.
Notably, the old function conflated support with "support enabled by running
`FloatNormalization`". Therefore, this change appears to impact what can be
fused end-to-end, but that should not be the case provided `FloatNormalization`
is run appropriately.

Since I had to fix the tests in this direction, I took the opportunity to
edit tests in order to remain relevant, and to delete the ones that no longer
seemed useful. All in all, the tests now focus on the properties of the matcher
itself, instead of also testing Triton support which now has its own
disentangled logic. (Except for a couple of dedicated tests that ensure that
Triton support is indeed taken into account when deciding to fuse a diamond.)

PiperOrigin-RevId: 649644761