[XLA:GPU] Add num_blocks to TiledHloComputation.
The original plan was to use `block_id_to_tile_offsets_indexing` to get this information, but turned out that it's too expensive to compute IndexingMap for each tiled instruction. This is a step to minimize the places where we compute the indexing.

The only case when number of blocks per tile is different is if the have a concat in the fusion. This case would have been supported by ``block_id_to_tile_offsets_indexing`, but now we need a new plan. Since `SymbolicTileAnalysis` also doesn't support `concat` yet, it's not an immediate issue.

PiperOrigin-RevId: 649650457