Move the utility code in the tf_saved_model_freeze_variables to a reusable place.
Get feature parity across freeze_global_tensors.cc and tf_saved_model_freeze_variables.cc, by reusing the underlying code that does the resource-propagation and freezing.

Both the passes/functions are trying to do similar things, one on an immutable GlobalTensorOp and the other on an immutable tf.VarHandleOp. So, it should make sense to reuse code to improve support.

freeze_global_tensors.cc appears to have issues, reported in the attached bug and currently only supports if the user of the resource is `TF::ReadVariableOp` or a `CallOpInterface`. But `tf_saved_model_freeze_variables.cc` has a broader support.

PiperOrigin-RevId: 650028180