[XLA:MSA] Implement an auxiliary function (SimulateAsyncCopyDone) to simulate the overhead of processing copy-done instruction.
The off-the-shelf runtime predictor does not include the overhead of asynchronous copies. We need to simulate the memory system to get the overhead of these copies. This CL implements a function to simulate the overhead of copy-done instructions. Specifically, there are two directions which share the bandwidth: default-read and default-write. Two directions will share the bandwidth equally. For example, when we process a default-read request, if there are also outstanding default-write process, we can only use half of the full bandwidth to process requests in each direction.

PiperOrigin-RevId: 652635815