[XLA:CPU] Add support for `transpose` to thunk runtime.
In old runtime, if `transpose` is not rewritten by any HLO pass, it is handled by the elemental generator. This commit introduces the same behavior to thunks runtime, also adds test case verifying that case (was missing).

Thunk runtime already supports all other ops to which transpose is rewritten, so no further changes are required. Turned on `transpose` tests for thunks runtime.

PiperOrigin-RevId: 653606107