Fix crash in AllReduceBlueConnect when multiple partitions are used.
Also, the pass now only runs when the all-reduce op has specific values for CollectiveOpGroupMode: kCrossReplica and kFlattenedID. Previously, the pass crashed with any mode other than kCrossReplica. I'm not sure when the two other modes, kCrossPartition and kCrossReplicaAndPartition, are used in JAX programs, and am unsure how to create HLO which uses these modes, so I decided not to support them for now.

PiperOrigin-RevId: 653840153