[XLA:GPU] Only compute tile offset indexing maps for instructions when it is needed.
Computing tile offset indexing maps is very expensive and not always necessary. When we're at Cost Model/Tiling stage, we only use the indexing map to deduplicate instruction. We can predict if we'll need an indexing map for a particular instruction by computation and comparing a parts of the hash.

PiperOrigin-RevId: 653988963