Handle atomic dumping of per fusion autotuning cache to cns.
To avoid multiple processes are writing to the same file, the per-fusion autotuning cache dumps the result to temporary file and then renames it, moving it to the location specified with the flag. This is not working out of the box for cns. Added support for this. The tmp cns folder should be passed by setting TMP envvar.

PiperOrigin-RevId: 665326014