[XLA:CPU] Add support for attributes in FFI instantiate callback.
This is the next step in stateful FFI support feature. Now all custom call attributes are available in the instantiate phase, in addition to the execution phase.

Note that there is no distinction between instantiate phase attributes and execution time attributes, so even if an attribute is only needed in one of these phases, it is still available in both of them.

PiperOrigin-RevId: 667577694