[XLA:GPU] Avoid applying a mask for the full tiles.
This is a performance optimization that can be applied to all Triton fusions.
When the contracting_dimension % block_k is non zero we need to apply the mask.
We do that unconditionally but it make sense to do that only for the tiles that are on the border of dim_k.

Let's check that the tile does not need masking and skip these ops.

PiperOrigin-RevId: 669225408