Store runtime and driver version as SemanticVersion in DeviceDescription
This allows replacing a whole bunch of usages of `CUDA_VERSION` and `TF_ROCM_VERSION` in tests by runtime conditionals.

`DeviceDescription::runtime_version` and `DeviceDescription::driver_version`
exist as a vendor dependent string which is only useful for displaying it to the user.

So this change:
1. Renames `DeviceDescription::runtime_version` to `DeviceDescription::runtime_version_string` (the same for the driver version)
2. Introduces new fields `DeviceDescription::runtime_version` and `::driver_version` of type `SemanticVersion`
3. Adds version parsers for CUDA and ROCm and makes the respective executors set the correct runtime and driver version.
4. Fixes up usagesof the old `runtime_version` and `driver_version` fields.

Replacing the macros in tests will follow in a subsequent change.

PiperOrigin-RevId: 670257004