[IFRT] Verify that devices is not nullptr when used in Sharding
`Sharding` currently expects one or more devices, and this expects any
`tsl::RCReference<DeviceList>` to be not `nullptr`. While this invariant may
change in the future depending on how we express an empty device list, we add a
check (in addition to `devices` access) that clarifies `devices != nullptr`
when used in `Sharding`.

PiperOrigin-RevId: 672620061