Make TritonSupportTest surface sanitizer issues
TritonSupportTest relies on a death test for some of its assertion because Triton doesn't allow graceful error handling for some of the cases we want to test.

This death test also used to succeed when the code under test triggered a santizer violation.

So this change makes it fail on those sanitizer violations and will be surface those in the log.

PiperOrigin-RevId: 693362123