Add `AutotuneCacheMode` to `xla_client.pyi`.
An `AutotuneCacheMode` enum was added to `xla_extension` and `xla_client` in https://github.com/openxla/xla/pull/18450, but it looks like it was missed in `xla_client.pyi`. This is one of the issues blocking the merge of https://github.com/jax-ml/jax/pull/22899, but I think this should do the trick!

PiperOrigin-RevId: 693649009