[XLA:SPMD] Allow `reshard_lhs_rhs_to_match_output_sharding` in the inner iterations of PartitionDot.
We process the dot operation recursively but only allow reshard_lhs_rhs_to_match_output_sharding in the most outer iteration. This cl enables it in the inner iterations.

For the added example
```
i is LHS non-contracting, j is contracting, k is RHS non-contracting, "a", "b", "c" are sharding axes
          i    j    k
LHS      "a"
RHS                "b"
Result   "a"       "b","c"
```
Previously, we (1) replicated RHS, (2) applied the dot, (3) partitioned the dot result along k, since we did not reshard_lhs_rhs_to_match_output_sharding in the inner iteration.

With this cl, we (1) partition RHS along k, and (2) applied the dot.

PiperOrigin-RevId: 694295034