The current check is overstringent for the following case:
Each slide is of shape (2, 1, 2),
the origins in lexicographical order are (0, 0, 0), (0, 0, 2), (0, 0, 4), (2, 0, 0), ...

The offset between the third and 4th origins are (2, 0, 4)
and the offset on the last dimenion 4 is larger than 2, causing
the check to false alarm.

Enforcing this check seems non-trivial so this CL simply
relax the check by ignoring offset difference when
the outer dimension in lexicographical order changes.

PiperOrigin-RevId: 706135726