PR #21437: [ds-fusion] Fix algebraic simplifier error in debug mode.
Imported from GitHub PR https://github.com/openxla/xla/pull/21437

This error was observed while trying to land #21375 (which is needed for the ds-fusion work).

This error occurs when there is a constant operation that can be converted into a scalar broadcast, but some other operation is a successor for the constant operation (via control dependency). Such a dependency is not relayed and so the operation is not converted even after the
`ReplaceWithNewInstruction` function call. This causes a runtime error in debug mode testing. Fixing this by relaying this control dependency.
Copybara import of the project:

--
0e869b268215a4be61e94eca0a62c57e6f4471cb by Shraiysh Vaishay <svaishay@nvidia.com>:

[ds-fusion] Fix algebraic simplifier error in debug mode.

This error was observed while trying to land #21375 (which is needed for
the ds-fusion work).

This error occurs when there is a constant operation that can be converted into a scalar
broadcast, but some other operation is a successor for the constant
operation (via control dependency). Such a dependency is not relayed and
so the operation is not converted even after the
`ReplaceWithNewInstruction` function call. This causes a runtime error
in debug mode testing. Fixing this by relaying this control dependency.

Merging this change closes #21437

FUTURE_COPYBARA_INTEGRATE_REVIEW=https://github.com/openxla/xla/pull/21437 from shraiysh:algebraic_simplifier_fix 0e869b268215a4be61e94eca0a62c57e6f4471cb
PiperOrigin-RevId: 715658783