In literal_util.cc, use absl::uniform_int_distribution.
absl::uniform_int_distribution is faster than std::uniform_int_distribution. This makes initializing literals in run_hlo_module faster. In particular, I tested the following HLO:

    ENTRY f {
      arg = s8[2000000000] parameter(0)
      ROOT add_result = s8[2000000000] add(arg, arg)
    }

It takes 7.8 seconds to initialize the input literal with the absl function, and 18.2 with the std function.

Unfortunately the absl version of uniform_real_distribution is not faster. It takes 25.5 seconds with absl and 8.3 with std on the HLO when s8 is replaced with f16.

PiperOrigin-RevId: 723316656