[XLA] Preserve AUTO layout when converting from HLO to StableHLO
In HLO, AUTO layout is encoded as missing layout in `entry_computation_layout`.

In StableHLO, it's marked using `mhlo.layout_mode = "auto"` attribute of the main@ function argument or return value.

The behavior is off by default everywhere except the "hlo-translate" tool.

In the next change, I'll switch it on by default, except for places where it causes problems.

PiperOrigin-RevId: 723489371