Add subgraph_input_names(), subgraph_output_names() to the SignatureRunner
In LiteRT, it assumes that the order of names are aligned with Subgraph.
But the existing input_names(), output_names() API doesn't follow it, also
there are customers who rely on the legacy order.

This cl creates these new methods which returns the names which reflects
the underlying Subgraph I/O order.

PiperOrigin-RevId: 726116089