[XLA:GPU][Emitters] Restrict the inliner.
At the moment we inline a callee that calls one of the functions called by the caller. This PR adjusts the logic to inline only such callees that call a subset of the caller's functions. That way we can be sure that after inlining the caller calls the same set of functions excluding the inlined one.

Background: https://github.com/jax-ml/jax/issues/26162 contains an example of a MoF fusion that takes forever to compile.

The [indexing-based partitioner](https://github.com/openxla/xla/commit/44bc81687d83a328b9212f6ecf3e6c18a79bd5fd) in combination with this change fix the issue.

PiperOrigin-RevId: 729079659