Switch xla_test, etc to static linking within Google.
Previously, we switched xla_test, etc to static linking to catch duplicate main() definitions at build time. We had to revert the change as it increased test binary sizes and broke Nvidia's build.

In this second attempt, we make the change only for the Google internal build, so that external users aren't affected.

PiperOrigin-RevId: 729694439