[XLA:Collective] Get participant ids, count and group sizes from CollectiveOpGroupMode.
It is not very obvious from the code or comments, but we can use `GetReplicaGroupCountAndSize` and `GetParticipatingFlattenedIdGroups` only for `AllReduce` and `AllGather`, because of the check in  `GetCollectiveUseGlobalDeviceIds`. There is no reason why `CollectiveOpGroupMode` of other instruction would not be enough to get the information.

The check of `AllReduce` and `AllGather` was added in https://github.com/openxla/xla/pull/20024, but I couldn't find explanation why it wouldn't work of other instruction.

PiperOrigin-RevId: 739163964