[XLA:GPU] update ncu-rep parser to work with multiple kernels
Now the correctly processes profiles with multiple kernels as a list of rows (previously replaced kernels with the same name making analysis of multiple kernels impossible).
To enable processing of complex profiles we now have aggregation and filter. Aggregation looks at the prefix of the metric and might not be completely correct at the moment.

Filtering allows to pick a subset of kernels and only aggregate ones that we care about, e.g. --filter='after:name:setup' will only consider kernels that were run after the kernel with "setup" in the name. Again filter is somewhat simplistic but overall that provides an ability to e.g. sum durations or memory transfer across multiple kernels that nsight compute does not provide.

Misc:

- Updated argument names, profile name is passed as an argument, making the tool easier to use.

- To properly aggregate values it now always works with basic units, for example 'ns' instead of 's' or 'ms'.

PiperOrigin-RevId: 744698701