[XLA:GPU] Use padded tile sizes when computing launch configs.
While directly using the provided tile sizes is currently not a problem in
end2end codegen, it is risky to assume that the passed tile sizes are either
capturing the full dimension or are already powers of 2. In fact, we already
had a test case that didn't follow this constraint and also shows that the
wrong number of blocks was computed.

PiperOrigin-RevId: 746416922