Do not crash on autotuner verification errors in int4 tests
This is the only test we have in the codebase that considers this an error.
I don't think the test should demand this, since the fact that some configs produce wrong results looks completely orthogonal to int4 tests (and specifically subchanel dequantization cl/729199512, where this was added).

Yes, that's an issue, but an issue that we should tackle separately, and not in this test, since from the context of correctly fusing int4 into matmuls, this is a completely recoverable error.

This flag being set is causing a crash when we run with `--xla_gpu_experimental_enable_dynamic_dot_search_space` enabled.

PiperOrigin-RevId: 750983342