Refactor BFS function to accept distance parameter
This change adds an additional "distance" parameter to the BFS per_node_fn and expand_node_fn. The distance parameter indicates the distance of the node from the BFS starting nodes in the graph.
This parameter is used in GumgraphMatcher to set a minimum height for the subgraph when searching for mapping candidates, which should ensure the best candidate being considered in most cases.

Bottom-up matcher explores a minimum depth of one node layers during both candidate search and heuristic similarity calculation.

PiperOrigin-RevId: 752911415