[xla::ifrt] Rename xla::ifrt::Compiler::Compile to xla::ifrt::Compiler::CompileAndLoad and point xla::ifrt::Compiler::Compile to xla::ifrt::Compiler::CompileAndLoad.
The first of a number of CLs to build on pjrt's work splitting Compile into Compile and Load steps for supporting plugins.

For now, both Compile and CompileAndLoad at the ifrt level return LoadedExecutables. Once we update all uses of Compile, we will update Compile to map to pjrt's Compile, if available, which returns Executable.

PiperOrigin-RevId: 753285251