Load PyInfo from rules_python
This is to facilitate upgrading of rules_python and other parts of Bazel.

More specifically, this makes the code compatible with both PyInfo from
rules_python and the Bazel builtin PyInfo. It does this by making rules
accept and produce both providers. This allows the code to be compatible
with a mixture of Bazel builtin `py_*` rules and rules_python `py_*` rules,
which make upgrading rules_python and Bazel easier in the future.

Additional logic is present to handle two cases that will occur as part
of upgrading to Bazel 8 and higher:

* The builtin PyInfo can be an alias to rules_python PyInfo. This occurs
  when Bazel's autoloading is enabled. Because accepting/producing the
  same provider is an error, additional logic is needed to only use one.
* The builtin PyInfo can be None or a stub provider. This occurs occurs
  when Bazel's autoloading is disabled and builtin providers are also
  disabled. Because None or a stub provider would be an error, additional
  logic is needed to use only one.

PiperOrigin-RevId: 760719992