Fix variable ops in the flex delegate
The flex delegate stores `tensorflow::Tensor*` values in resource tensors, which is not really correct, and there is no way to tell whether a resource tensor contains a `tensorflow::Tensor*`, or an `int32_t` from a standard kTfLiteResource tensor.

The solution used in this change is a total hack: always allocate 8 bytes for `tensorflow::Tensor*` (even on 32-bit systems), so we can tell if the tensor contains an `int32_t` or `tensorflow::Tensor*`.

A few other alternatives considered and ruled out:
- Adding a new `kTfLiteTensorflowResource` tensor type to distinguish from `kTfLiteResource`.
- Storing these in `kTfLiteVariant` instead.

Both of these require changing the type of the tensor, which also seems hacky and may break other things in the future too. For this to work properly, the TFlite graph authoring tool would have to know which resource tensors would be handled by the flex delegate (to give them the appropriate type), which doesn't make sense (no TFlite model should make assumptions about what delegates will be used).

PiperOrigin-RevId: 768346773