[xla:layout_assignment] Support buffer types.
Change Shape and ShapeUtil to support the query and modification to the layout of the array storage inside a buffer type.

Enhance layout assignment to support buffer types. There are still problems, such as supporting operand_layout_constraints in layout assignment, which will be addressed in CLs in the sequence.

PiperOrigin-RevId: 775677425