Fix invalid process count bug?
We were seeing `StreamExecutorGpuClient::GetLatestIncarnations` erroring in
`std::vector::__throw_length_error`. Looking through the code, I have a hunch
that `topology_.ProcessCount()` was returning -1, and we were trying to create
a vector of size -1.

Digging through the code, it seems like if the topology is not "symmetric",
then `ProcessCount()` returns -1. The code says a job is not symmetric if
different slices have different numbers of hosts or different hosts have
different numbers of devices.

This commit passes the number of nodes directly to the
`StreamExecutorGpuClient`, hopefully fixing the problem.

PiperOrigin-RevId: 776628588