Speed up LatencyHidingScheduler by improving data structures in various ways:
(1) HloGraphNode had three InlinedVector<int64_t, 1> fields
(released_non_extendable_resources_, released_shareable_resources_,
occupied_shareable_resources_), an absl::flat_hash_map<int64_t,
int64_t> recursive_resources_, and a ResourcesVector resources_ (a
std:vector<std::pair<int64_t, int64_t>) that were almost all always
empty (99% of time in one large compilation).  We now store these in
an HloGraphNode::Rare struct that we allocate lazily in the
HloScheduleGraph parent object, and keep a pointer to the appropriate
entry if any of the objects are non-empty.  We allocate a single
shared Rare object with all empty fields that is pointed to by the
vast majority of HloGraphNode objects.

(2) Rather than having two std::vector<HloEdge> objects in each
HloGraphNode to hold predecessors and successors, we now have two
shared std::vector<HloEdge> in the HloScheduleGraph object to hold
all predecessors and successors across all HloGraphNodes, with contiguous
ranges in these vectors pointed to by individual HloGraphNode for
storing their HloEdge objects.

(3) Various boolean fields turned into single-bit bitfields

(4) Various int64_t fields turned into int32_t

(5) HloOpcode for each HloGraphNode now stored in the HloGraphNode itself,
rather than forcing reading it through the HloInstruction* object.  This
can sometimes lead to one less level of indirection when scheduler code
has an HloGraphNode in its hand.

As a result of the changes above, sizeof(HloGraphNode) is now 120 bytes instead
of 288 bytes.

For HloGraphNode memory, previously we allocated these individually via a
make_unique<HloGraphNode> that was the value in a
absl::flat_hash_map<const HloInstruction*, std::unique_ptr<HloGraphNode>>.
Instead of this, we now allocate a std::vector<HloGraphNode> in the HloScheduleGraph
object, so that all the storage for HloGraphNode objects is contiguous, and
we keep an
absl::flat_hash_map<const HloInstruction*, int>, where the value is the index
in the std::vector<HloGraphNode> objects.

Also improved the way in which instructions are compared and results
of the comparison propagated.  Previously, the signature of the
comparison function was:

```
  DefaultSchedulerCore::CandidateResult ReadySetLt::operator()(
    DefaultSchedulerCore::ScheduleCandidate& a,
    DefaultSchedulerCore::ScheduleCandidate& b) const;
```

Where CandidateResult was a struct containing a reference to a
ScheduleCandidate (one of a or b), and a const char* reason.

The new code does away with the CandidateResult object entirely, and
has the following signature and semantics:

```
  // "a" is a candidate instruction, and "b" is the best instruction found so
  // far.  Compare "a" to "b" to determine if "a" should replace "b" as a better
  // scheduling candidate.  If "a" (the candidate) is "better" than b (the best
  // node found so far), it returns true and stores "a" in "b". Otherwise it
  // returns false. *reason is always update to point to a string describing
  // which heuristic ultimately made the decision.
  bool ReadySetLt::MaybeUpdate(DefaultSchedulerCore::ScheduleCandidate& a,
                   DefaultSchedulerCore::ScheduleCandidate& b,
                   const char** reason) const;
```

This enables the result to be returned in a register, and for the comparison
code to update b directly, rather than going indirectly through the CandidateResult.

ScheduleCandidate was also reworked to use explicit has_... fields, rather than
using multiple std::optional<T> fields.  The problem with the latter is that it
introduces multiple conditional branches when copying a ScheduleCandidate object,
rather than just a simple memcpy of sizeof(ScheduleCandidate).

PiperOrigin-RevId: 786704510