[XLA:MSA] Speed up FixSchedule by iterating over instructions once.
Currently, in `FixSchedule` for each computation we construct the schedule sequence by iterating through all instruction indices, first adding the before instructions, then the current instruction and finally the after instructions whose `parent=computation`.

This is quite wasteful as we could be iterating over instructions whose parent does not match the current computation (resulting in no-ops). In this CL, we iterate once over the instruction indices (in order) and construct the sequences for all computations in parallel.

PiperOrigin-RevId: 789696241