Replace AffineMap by SymbolicMap inside IndexingMap
This CL migrates indexing_map.cc from using mlir::AffineExpr to the new xla::SymbolicExpr.

This change is the CL to replace the MLIR Affine dialect components with the custom SymbolicExpr abstraction within the XLA HLO analysis. SymbolicExpr provides a more tailored API for expression manipulation and analysis in the context of XLA.

Key changes in this CL include:

Replacing mlir::AffineExpr and related MLIR types with xla::SymbolicExpr and xla::SymbolicExprType.
Updating helper functions and the SymbolicExprSimplifier class to operate on SymbolicExpr.
Adapting the IndexingMap constructors and methods to use SymbolicMap and SymbolicExpr.
Introducing shims and TODOs (linked to b/446858351 and b/446856820) to maintain compatibility with code still using mlir::AffineMap during the transition.
Updating tests to reflect the changes in IndexingMap.
The affine_map_ member in IndexingMap is now a mutable cache to support gradual migration, and it's invalidated when the underlying symbolic_map_ changes.

Many CLs has been previously merged, so this migrations do not break any test neither cause a regression.

PiperOrigin-RevId: 802100018