[XLA:GPU]: Add triton_xla.atomic_spin_wait op
atomic_spin_wait takes in a pointer to an address or a tensor of addresses,
atomically loads the value of that address with the passed memory semantic and
scope, and compares it against the given value based on the comparator.

As long as the comparison returns true the operation will spin wait.

PiperOrigin-RevId: 802547402