Enable hermetic toolchain support in configure.py
This change allows `configure.py` to support builds using a hermetic Clang toolchain on Linux x86_64 and aarch64. When a hermetic build is detected, the script will not search for system-installed `clang` and `lld`. The generated bazelrc will only include paths to Clang if a path was explicitly provided or found on the system.

PiperOrigin-RevId: 802895976