Adds `TfrtGpuClient::CreateAliasBuffer` to create a placeholder buffer.
This method creates a `PjRtBuffer` and a callback. The returned buffer acts as an alias, and its underlying memory and definition event are linked to a real `PjRtBuffer` when the provided callback is executed. The callback takes the real `PjRtBuffer` as input and forwards the alias buffer's internal `AsyncValue` and definition event to those of the real buffer.

PiperOrigin-RevId: 807853612