Don't construct GPU cliques with stale incarnations.
Recall that every process in a multi-controller JAX program is given a global
unique incarnation id. If the task fails and restarts, it is given a new
incarnation id.

Previously, when we detected the failure of a process with incarnation id x, we
aborted all collectives where process x was a participant. In other words, if a
task failed after forming a clique, things worked great. However, we didn't
handle the case where a task fails before forming a clique. This is illustrated
by the following scenario:

- live_devices is called and returns incarnations x, y and z.
- x fails.
- All cliques with x are aborted, but there are none.
- A clique is formed with incarnations x, y, and z.
- Things get stuck :(

This commit tracks the latest set of incarnations. If any of these incarnations
dies, the corresponding cliques are aborted. And, if we attempt to construct a
clique with a stale incarnation, it fails.

PiperOrigin-RevId: 810623801