[XLA] Add kScratch/kConsume to possible buffer uses
kScratch is a special case of kWrite, where the buffer contents are undefined
after use. It's mostly useful in combination with kRead, to allow
distinguishing buffers that are read from and keep their contents during thunk
execution from ones that are used as scratch space and can't be relied upon
afterwards.

PiperOrigin-RevId: 814233688