[IFRT] Introduce `ShardingSpec` and subclasses
`ShardingSpec` and its subclasses are a device-agnostic alternative to `Sharding` and its subclasses. Instead of using a concrete device or device list, it only tracks the number of devices in a partitioning scheme. This makes it easy to serialize/deserialize it when a device assignment may change (e.g., model export/import roundtrip) and when carrying the partitioning information to a non-IFRT-client context where concrete IFRT devices are undefined (e.g., non-client internals of transports).

`Sharding` and subclasses untouched in this CL. They will eventually be migrated to use `ShardingSpec` and subclasses internally, essentially becoming a tuple of (`ShardingSpecRef`, `DeviceListRef`, and `MemoryKind`), plus a few helper methods.

PiperOrigin-RevId: 846367600