Enforce that TFRT GPU does not call CUDA APIs inline
This CL implements a new guarantee in TFRT GPU where it never calls CUDA APIs on the same thread as the caller of PjRt APIs and fixes all call sites that violate this property. This property is useful because of the following reasons:

1. CUDA uses synchronization primitives that are not compatible with cooperatively scheduled threads. For example, Google's fiber scheduler are not aware of blocking code inside CUDA, which results in suboptimal scheduling decisions.
2. TFRT GPU's goal is to perform GPU operations asynchronously. If CUDA API is being called inline, this is likely an indication that the work is not being performed asynchronously by accident.

This new property is verified by a thread-local variable inside `TfrtGpuThreadChecker` that is set at the beginning of every unit test and checked upon every `se::Stream` and `se::StreamExecutor` access. This seems to be the least intrusive way to verify the property since all CUDA APIs are encapsulated behind the StreamExecutor API and unit tests are supposed to cover all public PjRt APIs.

Note that this property is satisfied only if TFRT GPU client is created with `GpuAllocatorConfig::preallocate = true`. This is because without preallocation, TFRT GPU may call `cuMemAlloc`/`cuMemFree` inline any time the allocator runs out of memory. While we may introduce an additional thread hop per `Allocate` call, that seems wasteful given that most users use `preallocate = true` anyway.

PiperOrigin-RevId: 850126895