[XLA] Limit 5cccc03 only to instructions with control dependencies
Removing an instruction from the schedule and then rescheduling it causes it and all of its successors to be scheduled in a post-order fashion, effectively destroying the schedule if it was non-trivial.

Ideally HloSchedule::UpdateSchedule() should handle control successors better, but for now limit the scope of the original workaround.

PiperOrigin-RevId: 850188449